Allow optional inputs to be Input<T | undefined>
The Node.js runtime accepts `Output<undefined>` values as inputs to optional parameters, but the TypeScript typing currently does not allow these.  This extends the TypeScript typings to allow `Output<undefined>` values as inputs to optional input properties.

I cannot think of any way in which this is breaking or would regress any aspect of the TypeScript experience, other than making the `.d.ts` files a little "noisier".

Fixes #6175